% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrivingFactors.R
\name{Decomposition}
\alias{Decomposition}
\title{Decomposition of attributed deaths}
\usage{
Decomposition(serie, G, D_r, D_c = NULL, P, A, M, L = NULL, key, from, to)
}
\arguments{
\item{serie}{# Calc Steps for Each Series:
1. PG-PA-EXP-ORF
2. PG-PA-ORF-EXP
3. PG-EXP-PA-ORF
4. PG-EXP-ORF-PA
5. PG-ORF-PA-EXP
6. PG-ORF-EXP-PA
7. PA-PG-EXP-ORF
8. PA-PG-ORF-EXP
9. PA-EXP-PG-ORF
10. PA-EXP-ORF-PG
11. PA-ORF-PG-EXP
12. PA-ORF-EXP-PG
13. EXP-PG-PA-ORF
14. EXP-PG-ORF-PA
15. EXP-PA-PG-ORF
16. EXP-PA-ORF-PG
17. EXP-ORF-PA-PG
18. EXP-ORF-PG-PA
19. ORF-PG-PA-EXP
20. ORF-PG-EXP-PA
21. ORF-PA-PG-EXP
22. ORF-PA-EXP-PG
23. ORF-EXP-PG-PA
24. ORF-EXP-PA-PG}

\item{G}{refers to `field` in `Mortality()`}

\item{D_r}{refers to `dose_real` in `Mortality()`}

\item{D_c}{refers to `dose_cf` in `Mortality()`}

\item{P}{refers to `pop` in `Mortality()`}

\item{A}{refers to `age` in `Mortality()`}

\item{M}{refers to `mort` in `Mortality()`}

\item{L}{refers to `lvl` in `Mortality()`}

\item{key}{name of key storing the scenario/year information}

\item{from}{scenario/year when the driving space start}

\item{to}{scenario/year when the driving space end}
}
\value{
a data.frame of decomposed attributed deaths
}
\description{
Decomposition of attributed deaths
}
\examples{
1:24 \%>\% map(~ Decomposition(serie = .x, start.y = 'base2015', end.y = 'SSP1-Baseline_2030'))
}
