% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{GEMM_Lookup_Table}
\alias{GEMM_Lookup_Table}
\title{the GEMM C-R model lookuptable.}
\format{
data.frame
}
\usage{
GEMM_Lookup_Table
}
\description{
the GEMM C-R model lookuptable.
}
\keyword{datasets}
