% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{get_at}
\alias{get_at}
\title{filter data at specific scenario/year}
\usage{
get_at(x, where = scenario, at)
}
\arguments{
\item{x}{the input data}

\item{where}{specify the scenario/year column}

\item{at}{the year/scenario to choose}
}
\value{
data.frame, contains joining keys and the corrosponding data
}
\description{
filter data at specific scenario/year
}
