% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{IER2013_Lookup_Table}
\alias{IER2013_Lookup_Table}
\title{the IER C-R curves by GBD2013.}
\format{
list
}
\usage{
IER2013_Lookup_Table
}
\description{
the IER C-R curves by GBD2013.
}
\keyword{datasets}
