% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{IER2015_Lookup_Table}
\alias{IER2015_Lookup_Table}
\title{the IER2015 C-R model lookuptable.}
\format{
data.frame
}
\usage{
IER2015_Lookup_Table
}
\description{
the IER2015 C-R model lookuptable.
}
\keyword{datasets}
