% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{matchable}
\alias{matchable}
\title{format numbers to a string at a specified digit}
\usage{
matchable(num, dgt = 2)
}
\arguments{
\item{num}{input number}

\item{dgt}{int, refers to rounding digit}
}
\value{
string of number
}
\description{
format numbers to a string at a specified digit
}
\examples{
matchable(3.33333, 1)
}
