% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{Mortality}
\alias{Mortality}
\title{Calculate girded PM2.5 attributed mortality}
\usage{
Mortality(
  field,
  dose_real,
  dose_cf = NULL,
  pop,
  age,
  mort,
  lvl = NULL,
  RR = "MEAN"
)
}
\arguments{
\item{field}{attribution, deciding the calculating field}

\item{dose_real}{does data, a data.frame with column(s) of joining key(s) and a column of dose data}

\item{dose_cf}{does data, a data.frame with column(s) of joining key(s) and a column of counter-fact dose data}

\item{pop}{population data, a data.frame with column(s) of joining key(s) and a column of population}

\item{age}{population age-structure, a data.frame with column(s) of joining key(s) and a column of population}

\item{mort}{mortality data, a data.frame with column(s) of joining key(s) and a column of population}

\item{lvl}{level of resolution, must corresponds to the `mort` data, if NULL, the calculations will be implement at no calibration on mort data.}

\item{RR}{param passed to `RR_std()`}
}
\value{
data.frame. death estimates for each endpoint & age-groups(columns) for every grids(rows)
}
\description{
Calculate girded PM2.5 attributed mortality
}
