% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{Mortality_Aggr}
\alias{Mortality_Aggr}
\title{Aggregated mortality calculation and attach uncertainty.}
\usage{
Mortality_Aggr(
  field,
  dose_real,
  dose_cf = NULL,
  pop,
  age,
  mort,
  doseRSME = 0,
  lvl = NULL,
  aggr_by = NULL,
  uncertain = F,
  write = T
)
}
\arguments{
\item{field}{attribution, deciding the calculating field}

\item{dose_real}{does data, a data.frame with column(s) of joining key(s) and a column of dose data}

\item{dose_cf}{does data, a data.frame with column(s) of joining key(s) and a column of counter-fact dose data}

\item{pop}{population data, a data.frame with column(s) of joining key(s) and a column of population}

\item{age}{population age-structure, a data.frame with column(s) of joining key(s) and a column of population}

\item{mort}{mortality data, a data.frame with column(s) of joining key(s) and a column of population}

\item{doseRSME}{uncertainty of dose data}

\item{lvl}{the aggregate level, corresponds to the field's keys, string.}

\item{aggr_by}{additional keys to aggregate, age (group) or endpoint}

\item{uncertain}{incorporate uncertainty calculation in aggregation}
}
\value{
the aggregated burden
}
\description{
Aggregated mortality calculation and attach uncertainty.
}
