% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{MRBRT2021_Lookup_Table}
\alias{MRBRT2021_Lookup_Table}
\title{the MRBRT C-R curves for PM25 by GBD2021.}
\format{
list
}
\usage{
MRBRT2021_Lookup_Table
}
\description{
the MRBRT C-R curves for PM25 by GBD2021.
}
\keyword{datasets}
