% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{MRBRT_Lookup_Table}
\alias{MRBRT_Lookup_Table}
\title{the MRBRT C-R model lookuptable.}
\format{
data.frame
}
\usage{
MRBRT_Lookup_Table
}
\description{
the MRBRT C-R model lookuptable.
}
\keyword{datasets}
