% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{NO2_CR_Lookup_Table}
\alias{NO2_CR_Lookup_Table}
\title{the NO2 C-R model lookuptable.}
\format{
data.frame
}
\usage{
NO2_CR_Lookup_Table
}
\description{
the NO2 C-R model lookuptable.
}
\keyword{datasets}
