% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{O3_CR_Lookup_Table}
\alias{O3_CR_Lookup_Table}
\title{the O3 C-R model lookuptable.}
\format{
data.frame
}
\usage{
O3_CR_Lookup_Table
}
\description{
the O3 C-R model lookuptable.
}
\keyword{datasets}
