% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{RR_std}
\alias{RR_std}
\title{standardize CR look-up table}
\usage{
RR_std(index = "MEAN")
}
\arguments{
\item{index}{string, specifying witch RR table to use, by default the "MEAN" RR.}
}
\value{
a formatted RR table
}
\description{
standardize CR look-up table
}
