% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{set_Model}
\alias{set_Model}
\title{set CRF for calculation}
\usage{
set_Model(Model)
}
\arguments{
\item{Model}{string, one of `IER`, `NCD+LRI`, `5COD`, `MRBRT`, `O3` or `NO2`}
}
\description{
set CRF for calculation
}
