% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{tell_Model}
\alias{tell_Model}
\title{detect CRF model name and generate a name for file output}
\usage{
tell_Model()
}
\value{
formatted string of CRF name
}
\description{
detect CRF model name and generate a name for file output
}
